feat: add python to supported cgi

+++ sources/Poll.cpp
@@ -6,7 +6,7 @@
 /*   By: abel-mak <abel-mak@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/11 15:48:46 by abel-mak          #+#    #+#             */
-/*   Updated: 2021/12/30 16:26:20 by abel-mak         ###   ########.fr       */
+/*   Updated: 2021/12/31 13:58:30 by abel-mak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,20 +33,6 @@ Poll::~Poll(void)
 	}
 }
 
-// std::vector<int> getAllPorts(const std::vector<ServerData> &data)
-//{
-//	size_t i;
-//	std::vector<int> allPorts;
-//
-//	i = 0;
-//	while (i < data.size())
-//	{
-//		allPorts.push_back(data[i].getPort());
-//		i++;
-//	}
-//	return (allPorts);
-// }
-
 /*
  * set up socket fds for each port
  * set socket as activefds so we can monitor them using select
@@ -56,7 +42,7 @@ Poll::Poll(std::vector<ServerData> data) : _queue(10), _data(data)
 {
 	int tmpMasterSocket;
 	int optval;
-	int i;
+	size_t i;
 	struct sockaddr_in server;
 
 	FD_ZERO(&_writeActivefds);
@@ -127,40 +113,19 @@ Poll::Poll(std::vector<ServerData> data) : _queue(10), _data(data)
 std::vector<int> Poll::getReadyfds(void)
 {
 	int sret;
-	int i;
+	size_t i;
 	int slavesocket;
 	std::vector<int> res;
 	std::map<int, std::string> _rawRequest;
 	std::string response;
 	unsigned int len;
-	// struct timeval tv;
-
-	// i = 0;
-	// FD_ZERO(&_writefds);
-	// while (i < _writeReadyFds.size())
-	//{
-	//	FD_SET(i, &_writefds);
-	//	i++;
-	// }
 
 	/**************************************************************************/
 
-	// tv.tv_sec  = 5;
-	// tv.tv_usec = 0;
-	//i = 0;
-	// while (i < FD_SETSIZE)
-	//{
-	//	if (FD_ISSET(i, &_activefds) != 0)
-	//	{
-	//		// std::cout << "active: " << i << std::endl;
-	//	}
-	//	i++;
-	// }
 	memcpy(&_writefds, &_writeActivefds, sizeof(_writefds));
 	memcpy(&_readfds, &_activefds, sizeof(_readfds));
-	// std::cout << "selec BEGIN" << std::endl;
+
 	sret = select(FD_SETSIZE, &_readfds, &_writefds, NULL, NULL);
-	// std::cout << "selec END" << std::endl;
 
 	i = 0;
 	while (i < _masterSockets.size())
